pyproject.toml
[tool.poetry]
name = "mqtt-zentrale"
version = "0.1.0"
description = ""
authors = ["Max Hoefl"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"
fastapi = "^0.104.1"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
